requirements.txt
langchain[google-genai]
streamlit < 1.33.99
streamlit-folium < 0.19.99
httpx==0.27.2
jupyter 
pandas
python-dotenv
langchain-community
# langchain-google-genai==2.0.5